+++ include/vml/allgen.sh
@@ -1,3 +1,4 @@
+#!/bin/bash
 sh ./gen_vector_vml.sh > ./monolish_vector_vml.hpp
 sh ./gen_dense_vml.sh > ./monolish_dense_vml.hpp
 sh ./gen_crs_vml.sh > ./monolish_crs_vml.hpp