+++ src/vml/matrix/math_1_3op/matpow.hpp
@@ -1,3 +1,5 @@
+#pragma once
+
 namespace monolish {
 namespace {
 template <typename F1, typename F2, typename F3>